requirements.txt
requests
aiofiles
aiohttp
vol
python-dateutil
